Remove tds subpackage and replace with catalog module.
