+++ siphon/__init__.py
@@ -1,8 +1,7 @@
 # use __init__.py to setup the namespace
-from . import tds
 from . import cdmr
 from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
 
-__all__ = ['tds', 'cdmr']
+__all__ = ['catalog', 'cdmr']
