--- siphon/tds/__init__.py
@@ -1,3 +1,0 @@
-from .TDSCatalog import TDSCatalog, get_latest_access_url
-
-__all__ = ['TDSCatalog', 'get_latest_access_url']
