--- siphon/tds/test/test_catalog.py
@@ -1,16 +1,0 @@
-from siphon.tds import TDSCatalog
-
-
-class TestCatalog(object):
-    baseURL = 'http://thredds-test.unidata.ucar.edu/thredds/'
-
-    def test_basic(self):
-        url = self.baseURL + 'catalog.xml'
-        cat = TDSCatalog(url)
-        assert 'Forecast Model Data' in cat.catalog_refs
-
-    def test_access(self):
-        url = self.baseURL + 'catalog/grib/NCEP/GFS/Global_0p5deg/latest.xml'
-        cat = TDSCatalog(url)
-        ds = list(cat.datasets.values())[0]
-        assert 'OPENDAP' in ds.access_urls
