--- siphon/tds/test/test_dods.py
@@ -1,11 +1,0 @@
-import os.path
-import vcr
-from siphon.tds import get_latest_access_url
-
-
-@vcr.use_cassette(os.path.join(os.path.dirname(__file__), 'fixtures/latest_rap_catalog'))
-def test_get_latest():
-    url = ('http://thredds-test.unidata.ucar.edu/thredds/catalog/'
-           'grib/NCEP/RAP/CONUS_13km/catalog.xml')
-    latest_url = get_latest_access_url(url, "OPENDAP")
-    assert latest_url
